backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
pytest==8.3.0
httpx==0.27.0